+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 0.24.1
 - Added ability to choose which example(s) to regenerate from the cli using the
   `--examples` flag. Defaults to animals and satellites.
+- Added ability to specify `n_iters` and `timeout` for examples re-generation
 
 ## 0.24.0
 - Implemented new data structure for columnar data in `braid_data` subcrate.